@@ -31,6 +31,15 @@ import {MatTableModule} from '@angular/material/table';
 import { TemplateEvaluateResultComponent } from './template-evaluate/template-evaluate-result/template-evaluate-result.component'; 
 import { TemplateEvaluateComponent} from './template-evaluate/template-evaluate.component'
 import { ModelNewComponent } from './model-new/model-new.component'; 
+import { ModelValidateComponent } from './model-validate/model-validate.component';
+import { LibraryService } from './library/library.service';
+import {MatSidenavModule} from '@angular/material/sidenav'; 
+import {MatTabsModule} from '@angular/material/tabs';
+import {MatCheckboxModule} from '@angular/material/checkbox'; 
+import {MatDialogModule} from '@angular/material/dialog';
+import {MatStepperModule} from '@angular/material/stepper';
+import { ShapeValidationComponent } from './shape-validation/shape-validation.component';
+import {MatExpansionModule} from '@angular/material/expansion'; 
 
 @NgModule({
   declarations: [
@@ -45,6 +54,8 @@ import { ModelNewComponent } from './model-new/model-new.component';
     TemplateEvaluateFormComponent,
     TemplateEvaluateResultComponent,
     ModelNewComponent,
+    ModelValidateComponent,
+    ShapeValidationComponent,
   ],
   imports: [
     BrowserModule,
@@ -66,9 +77,15 @@ import { ModelNewComponent } from './model-new/model-new.component';
     MatSnackBarModule,
     CodemirrorModule,
     MatCardModule,
-    MatTableModule
+    MatTableModule,
+    MatSidenavModule,
+    MatTabsModule,
+    MatCheckboxModule,
+    MatDialogModule,
+    MatStepperModule,
+    MatExpansionModule,
   ],
-  providers: [TemplateDetailService],
+  providers: [TemplateDetailService, LibraryService],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
