@@ -104,8 +104,8 @@ class BuildingMOTIF(metaclass=Singleton):
         engine_logger = logging.getLogger("sqlalchemy.engine")
         pool_logger = logging.getLogger("sqlalchemy.pool")
 
-        engine_logger.setLevel(logging.DEBUG)
-        pool_logger.setLevel(logging.DEBUG)
+        engine_logger.setLevel(logging.WARN)
+        pool_logger.setLevel(logging.WARN)
 
         stream_handler = logging.StreamHandler()
         stream_handler.setLevel(log_level)
