@@ -1,5 +1,6 @@
 import logging
 import uuid
+from functools import lru_cache
 from typing import Dict, List, Tuple
 
 from sqlalchemy.engine import Engine
@@ -12,7 +13,6 @@ from buildingmotif.database.tables import (
     DBTemplate,
     DepsAssociation,
 )
-from buildingmotif.utils import get_parameters
 
 
 class TableConnection:
@@ -41,11 +41,21 @@ class TableConnection:
         :return: DBModel
         :rtype: DBModel
         """
+        manifest_id = str(uuid.uuid4())
+        self.logger.debug(f"Creating shape collection in model: '{name}'")
+        manifest = DBShapeCollection(graph_id=manifest_id)
+        self.bm.session.add(manifest)
+
         graph_id = str(uuid.uuid4())
         self.logger.debug(f"Creating model: '{name}', with graph: '{graph_id}'")
-        db_model = DBModel(name=name, graph_id=graph_id, description=description)
-
+        db_model = DBModel(
+            name=name,
+            graph_id=graph_id,
+            description=description,
+            manifest=manifest,
+        )
         self.bm.session.add(db_model)
+
         self.bm.session.flush()
 
         return db_model
@@ -198,7 +208,7 @@ class TableConnection:
         db_libraries = self.bm.session.query(DBLibrary).all()
         return db_libraries
 
-    def get_db_library_by_id(self, id: int) -> DBLibrary:
+    def get_db_library(self, id: int) -> DBLibrary:
         """Get database library by id.
 
         :param id: id of DBLibrary
@@ -227,7 +237,7 @@ class TableConnection:
         :param name: new name
         :type name: str
         """
-        db_library = self.get_db_library_by_id(id)
+        db_library = self.get_db_library(id)
         self.logger.debug(
             f"Updating database library name: '{db_library.name}' -> '{name}'"
         )
@@ -240,7 +250,7 @@ class TableConnection:
         :type id: int
         """
 
-        db_library = self.get_db_library_by_id(id)
+        db_library = self.get_db_library(id)
 
         self.logger.debug(f"Deleting database library: '{db_library.name}'")
         self.bm.session.delete(db_library)
@@ -258,7 +268,7 @@ class TableConnection:
         :rtype: DBTemplate
         """
         self.logger.debug(f"Creating database template: '{name}'")
-        library = self.get_db_library_by_id(library_id)
+        library = self.get_db_library(library_id)
         template = DBTemplate(
             name=name,
             body_id=str(uuid.uuid4()),
@@ -280,7 +290,7 @@ class TableConnection:
         db_templates = self.bm.session.query(DBTemplate).all()
         return db_templates
 
-    def get_db_template_by_id(self, id: int) -> DBTemplate:
+    def get_db_template(self, id: int) -> DBTemplate:
         """Get database template by id.
 
         :param id: id of DBTemplate
@@ -306,7 +316,7 @@ class TableConnection:
                 self.bm.session.query(DBTemplate).filter(DBTemplate.name == name).one()
             )
         except NoResultFound:
-            raise NoResultFound(f"No tempalte found with name {name}")
+            raise NoResultFound(f"No template found with name {name}")
         return db_template
 
     def get_library_defining_db_template(self, id: int) -> DBLibrary:
@@ -317,7 +327,7 @@ class TableConnection:
         :return: DBLibrary
         :rtype: DBLibrary
         """
-        return self.get_db_template_by_id(id).library
+        return self.get_db_template(id).library
 
     def get_db_template_dependencies(self, id: int) -> Tuple[DepsAssociation, ...]:
         """Get a template's dependencies and its arguments.
@@ -346,7 +356,7 @@ class TableConnection:
         :param name: new name
         :type name: str
         """
-        db_template = self.get_db_template_by_id(id)
+        db_template = self.get_db_template(id)
         self.logger.debug(
             f"Updating database template name: '{db_template.name}' -> '{name}'"
         )
@@ -367,10 +377,21 @@ class TableConnection:
         )
         db_template.optional_args = optional_args
 
-    def add_template_dependency(
+    def add_template_dependency_preliminary(
         self, template_id: int, dependency_id: int, args: Dict[str, str]
     ):
-        """Create dependency between two templates.
+        """Creates a *preliminary* dependency between two templates. This dependency
+        is preliminary because the bindings between the dependent/dependency templates
+        are *not validated*. This serves to populate the directed acyclic graph of dependencies between templates
+        before the parameter bindings are checked. This ensures that all of the parameters for
+        a template *and those in its dependencies* can all be identified and used as part of
+        the parameter binding check. The upshot of this process is dependencies between two
+        templates can refer to parameters in a template *or its dependencies*. This is necessary
+        to support templates that contain many nested components that refer to each other (such
+        as the s223:mapsTo property).
+
+        **Important:** Be sure to run "check_all_template_dependencies" to ensure all of your templates.
+        will work as you expect!
 
         :param template_id: dependant template id
         :type template_id: int
@@ -380,44 +401,15 @@ class TableConnection:
         :type args: Dict[str, str]
         :raises ValueError: if all dependee required_params not in args
         :raises ValueError: if dependant and dependency template don't share a
-            library
         """
         self.logger.debug(
             f"Creating depencency from templates with ids: '{template_id}' and: '{dependency_id}'"
         )
-        templ = self.get_db_template_by_id(template_id)
-        graph = self.bm.graph_connection.get_graph(templ.body_id)
-        params = get_parameters(graph)
-        dep = self.get_db_template_by_id(dependency_id)
-
-        # check parameters are valid
-        if not set(args.values()).issubset(params):
-            raise ValueError(
-                f"In {templ.name} the values of the bindings to {dep.name} must correspond to the "
-                "parameters in the dependant template."
-                f"Dependency {dep.name} refers to params {set(args.values()).difference(params)} "
-                f"that do not appear in template {templ.name}"
-            )
-        # do the same for the dependency
-        graph = self.bm.graph_connection.get_graph(dep.body_id)
-        dep_params = get_parameters(graph)
-        if not set(args.keys()).issubset(dep_params):
-            raise ValueError(
-                f"In {templ.name} the keys of the bindings to {dep.name} must correspond to the "
-                "parameters in the template dependency"
-            )
-
-        # TODO: do we need this kind of check?
+        templ = self.get_db_template(template_id)
         if "name" not in args.keys():
             raise ValueError(
-                f"The name parameter is required for the dependency '{templ.name}'"
+                f"The name parameter is required for the dependency '{templ.name}'."
             )
-        if len(params) > 0 and args["name"] not in params:
-            raise ValueError(
-                "The name parameter of the dependency must be bound to a param in this template."
-                f"'name' was bound to {args['name']} but available params are {params}"
-            )
-
         # In the past we had a check here to make sure the two templates were in the same library.
         # This has been removed because it wasn't actually necessary, but we may add it back in
         # in the future.
@@ -430,8 +422,76 @@ class TableConnection:
         self.bm.session.add(relationship)
         self.bm.session.flush()
 
-    def remove_template_dependency(self, template_id: int, dependency_id: int):
-        """Remove dependency between two templates.
+    def check_all_template_dependencies(self):
+        """
+        Verifies that all template dependencies have valid references to the parameters
+        in the dependency or (recursively) its dependencies. Raises an exception if any
+        issues are found. Checks all templates in the database.
+        """
+        # TODO: maybe optimize this to only check recently added templates
+        # (i.e. maybe those added since last commit?)
+        for db_templ in self.get_all_db_templates():
+            for dep in self.get_db_template_dependencies(db_templ.id):
+                self.check_template_dependency_relationship(dep)
+
+    @lru_cache(maxsize=128)
+    def check_template_dependency_relationship(self, dep: DepsAssociation):
+        """Verify that the dependency between two templates is well-formed. This involves
+        a series of checks:
+        - existence of the dependent and dependency templates is performed during the
+        :py:method:`database.template_connection.`add_template_dependency_preliminary` method
+        - the args keys appear in the dependency, or recursively in a template that is a dependency
+        of the named dependency
+        - the args values appear in the dependent template
+        - there is a 'name' parameter in the dependent template
+
+        :param dep: the dependency object to check
+        :type dep: DepsAssociation
+        :raises ValueError: if all dependee required_params not in args
+        :raises ValueError: if dependant and dependency template don't share a
+            library
+        """
+        template_id = dep.dependant_id
+        dependency_id = dep.dependee_id
+        args = dep.args
+        self.logger.debug(
+            f"Creating depencency from templates with ids: '{template_id}' and: '{dependency_id}'"
+        )
+        from buildingmotif.dataclasses import Template
+
+        templ = Template.load(template_id)
+        params = templ.inline_dependencies().parameters
+        dep_templ = Template.load(dependency_id)
+        dep_params = dep_templ.inline_dependencies().parameters
+
+        # check parameters are valid
+        if not set(args.values()).issubset(params):
+            raise ValueError(
+                f"In {templ.name} the values of the bindings to {dep_templ.name} must correspond to the "
+                "parameters in the dependant template."
+                f"Dependency {dep_templ.name} refers to params {set(args.values()).difference(params)} "
+                f"that do not appear in template {templ.name} ({params})."
+            )
+        # do the same for the dependency
+        binding_params = set(args.keys())
+        if not binding_params.issubset(dep_params):
+            diff = binding_params.difference(dep_params)
+            raise ValueError(
+                f"In {templ.name} the keys of the bindings to {dep_templ.name} must correspond to the "
+                "parameters in the template dependency. "
+                f"The dependency binding uses params {diff} which don't appear in the dependency template. "
+                f"Available dependency parameters are {dep_params}"
+            )
+
+        # TODO: do we need this kind of check?
+        if len(params) > 0 and args["name"] not in params:
+            raise ValueError(
+                "The name parameter of the dependency must be bound to a param in this template."
+                f"'name' was bound to {args['name']} but available params are {params}"
+            )
+
+    def delete_template_dependency(self, template_id: int, dependency_id: int):
+        """Delete dependency between two templates.
 
         :param template_id: dependant template id
         :type template_id: int
@@ -460,7 +520,7 @@ class TableConnection:
         :param library_id: id of the new library
         :type library_id: int
         """
-        db_template = self.get_db_template_by_id(id)
+        db_template = self.get_db_template(id)
         self.logger.debug(
             f"Updating database template library: '{db_template.library_id}' -> '{library_id}'"  # noqa
         )
@@ -472,7 +532,7 @@ class TableConnection:
         :param id: id of deleted DBTemplate
         :type id: int
         """
-        db_template = self.get_db_template_by_id(id)
+        db_template = self.get_db_template(id)
         self.logger.debug(f"Deleting template: '{db_template.name}'")
 
         self.bm.session.delete(db_template)
