@@ -1,15 +1,22 @@
 from dataclasses import dataclass
 from typing import TYPE_CHECKING, Dict, List, Optional
 
-import pyshacl
 import rdflib
 import rfc3987
+from rdflib import URIRef
 
 from buildingmotif import get_building_motif
 from buildingmotif.dataclasses.shape_collection import ShapeCollection
 from buildingmotif.dataclasses.validation import ValidationContext
-from buildingmotif.namespaces import A
-from buildingmotif.utils import Triple, copy_graph, rewrite_shape_graph
+from buildingmotif.namespaces import OWL, A
+from buildingmotif.utils import (
+    Triple,
+    copy_graph,
+    rewrite_shape_graph,
+    shacl_inference,
+    shacl_validate,
+    skolemize_shapes,
+)
 
 if TYPE_CHECKING:
     from buildingmotif import BuildingMOTIF
@@ -32,6 +39,7 @@ class Model:
     _description: str
     graph: rdflib.Graph
     _bm: "BuildingMOTIF"
+    _manifest_id: int
 
     @classmethod
     def create(cls, name: str, description: str = "") -> "Model":
@@ -59,6 +67,7 @@ class Model:
             _description=db_model.description,
             graph=graph,
             _bm=bm,
+            _manifest_id=db_model.manifest_id,
         )
 
     @classmethod
@@ -88,6 +97,7 @@ class Model:
             _description=db_model.description,
             graph=graph,
             _bm=bm,
+            _manifest_id=db_model.manifest_id,
         )
 
     @property
@@ -133,14 +143,31 @@ class Model:
         """
         self.graph += graph
 
-    def validate(self, shape_collections: List[ShapeCollection]) -> "ValidationContext":
-        """Validates this model against the given ShapeCollections.
+    def validate(
+        self,
+        shape_collections: Optional[List[ShapeCollection]] = None,
+        error_on_missing_imports: bool = True,
+        engine: Optional[str] = "pyshacl",
+    ) -> "ValidationContext":
+        """Validates this model against the given list of ShapeCollections.
+        If no list is provided, the model will be validated against the model's "manifest".
+        If a list of shape collections is provided, the manifest will *not* be automatically
+        included in the set of shape collections.
 
         Loads all of the ShapeCollections into a single graph.
 
         :param shape_collections: a list of ShapeCollections against which the
-            graph should be validated
+            graph should be validated. If an empty list or None is provided, the
+            model will be validated against the model's manifest.
         :type shape_collections: List[ShapeCollection]
+        :param error_on_missing_imports: if True, raises an error if any of the dependency
+            ontologies are missing (i.e. they need to be loaded into BuildingMOTIF), defaults
+            to True
+        :type error_on_missing_imports: bool, optional
+        :param engine: the engine to use for validation. "pyshacl" or "topquadrant". Using topquadrant
+            requires Java to be installed on this machine, and the "topquadrant" feature on BuildingMOTIF,
+            defaults to "pyshacl"
+        :type engine: str, optional
         :return: An object containing useful properties/methods to deal with
             the validation results
         :rtype: ValidationContext
@@ -149,37 +176,49 @@ class Model:
         # but also want a report. Is this the base pySHACL report? Or a useful
         # transformation, like a list of deltas for potential fixes?
         shapeg = rdflib.Graph()
+        if shape_collections is None or len(shape_collections) == 0:
+            shape_collections = [self.get_manifest()]
         # aggregate shape graphs
         for sc in shape_collections:
-            # inline sh:node for interpretability
-            shapeg += sc.graph
+            shapeg += sc.resolve_imports(
+                error_on_missing_imports=error_on_missing_imports
+            ).graph
+        # inline sh:node for interpretability
         shapeg = rewrite_shape_graph(shapeg)
+
+        # remove imports from sg
+        shapeg.remove((None, OWL.imports, None))
+
+        # skolemize the shape graph so we have consistent identifiers across
+        # validation through the interpretation of the validation report
+        shapeg = skolemize_shapes(shapeg)
+
         # TODO: do we want to preserve the materialized triples added to data_graph via reasoning?
         data_graph = copy_graph(self.graph)
-        valid, report_g, report_str = pyshacl.validate(
-            data_graph,
-            shacl_graph=shapeg,
-            ont_graph=shapeg,
-            advanced=True,
-            js=True,
-            allow_warnings=True,
-            # inplace=True,
-        )
-        assert isinstance(report_g, rdflib.Graph)
+
+        # validate the data graph
+        valid, report_g, report_str = shacl_validate(data_graph, shapeg, engine=engine)
         return ValidationContext(
             shape_collections,
+            shapeg,
             valid,
             report_g,
             report_str,
             self,
         )
 
-    def compile(self, shape_collections: List["ShapeCollection"]):
+    def compile(
+        self, shape_collections: List["ShapeCollection"], engine: str = "pyshacl"
+    ):
         """Compile the graph of a model against a set of ShapeCollections.
 
         :param shape_collections: list of ShapeCollections to compile the model
             against
         :type shape_collections: List[ShapeCollection]
+        :param engine: the engine to use for validation. "pyshacl" or "topquadrant". Using topquadrant
+            requires Java to be installed on this machine, and the "topquadrant" feature on BuildingMOTIF,
+            defaults to "pyshacl"
+        :type engine: str
         :return: copy of model's graph that has been compiled against the
             ShapeCollections
         :rtype: Graph
@@ -188,41 +227,11 @@ class Model:
         for shape_collection in shape_collections:
             ontology_graph += shape_collection.graph
 
-        ontology_graph = ontology_graph.skolemize()
+        ontology_graph = skolemize_shapes(ontology_graph)
 
         model_graph = copy_graph(self.graph).skolemize()
 
-        # We use a fixed-point computation approach to 'compiling' RDF models.
-        # We accomlish this by keeping track of the size of the graph before and after
-        # the inference step. If the size of the graph changes, then we know that the
-        # inference has had some effect. We do this at most 3 times to avoid looping
-        # forever.
-        pre_compile_length = len(model_graph)  # type: ignore
-        pyshacl.validate(
-            data_graph=model_graph,
-            shacl_graph=ontology_graph,
-            ont_graph=ontology_graph,
-            advanced=True,
-            inplace=True,
-            js=True,
-        )
-        post_compile_length = len(model_graph)  # type: ignore
-
-        attempts = 3
-        while attempts > 0 and post_compile_length != pre_compile_length:
-            pre_compile_length = len(model_graph)  # type: ignore
-            pyshacl.validate(
-                data_graph=model_graph,
-                shacl_graph=ontology_graph,
-                ont_graph=ontology_graph,
-                advanced=True,
-                inplace=True,
-                js=True,
-            )
-            post_compile_length = len(model_graph)  # type: ignore
-            attempts -= 1
-        model_graph -= ontology_graph
-        return model_graph.de_skolemize()
+        return shacl_inference(model_graph, ontology_graph, engine)
 
     def test_model_against_shapes(
         self,
@@ -252,22 +261,34 @@ class Model:
         results = {}
 
         for shape_uri in shapes_to_test:
-            targets = model_graph.triples((None, A, target_class))
+            targets = model_graph.query(
+                f"""
+                PREFIX rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#>
+                PREFIX rdfs: <http://www.w3.org/2000/01/rdf-schema#>
+                SELECT ?target
+                WHERE {{
+                    ?target rdf:type/rdfs:subClassOf* <{target_class}>
+
+                }}
+            """
+            )
             temp_model_graph = copy_graph(model_graph)
-            for s, _, _ in targets:
-                temp_model_graph.add((s, A, shape_uri))
+            for (s,) in targets:
+                temp_model_graph.add((URIRef(s), A, shape_uri))
 
             temp_model_graph += ontology_graph.cbd(shape_uri)
 
-            valid, report_g, report_str = pyshacl.validate(
-                data_graph=temp_model_graph,
-                ont_graph=ontology_graph,
-                allow_warnings=True,
-                advanced=True,
-                js=True,
+            # skolemize the shape graph so we have consistent identifiers across
+            # validation through the interpretation of the validation report
+            ontology_graph = ontology_graph.skolemize()
+
+            valid, report_g, report_str = shacl_validate(
+                temp_model_graph, ontology_graph
             )
+
             results[shape_uri] = ValidationContext(
                 shape_collections,
+                ontology_graph,
                 valid,
                 report_g,
                 report_str,
@@ -275,3 +296,20 @@ class Model:
             )
 
         return results
+
+    def get_manifest(self) -> ShapeCollection:
+        """Get ShapeCollection from model.
+
+        :return: model's shape collection
+        :rtype: ShapeCollection
+        """
+        return ShapeCollection.load(self._manifest_id)
+
+    def update_manifest(self, manifest: ShapeCollection):
+        """Updates the manifest for this model by adding in the contents
+        of the shape graph inside the provided SHapeCollection
+
+        :param manifest: the ShapeCollection containing additional shapes against which to validate this model
+        :type manifest: ShapeCollection
+        """
+        self.get_manifest().graph += manifest.graph
