@@ -1,10 +1,15 @@
 from flask_api import status
 from rdflib import Graph, Namespace
 
-from buildingmotif.dataclasses import Library
+from buildingmotif.dataclasses import Library, Model
 from buildingmotif.namespaces import BRICK, A
 
 BLDG = Namespace("urn:building/")
+
+BINDINGS = b"""name,cav
+zone1,cav1
+zone2,cav2
+"""
 
 
 def test_get_all_templates(client, building_motif):
@@ -44,7 +49,7 @@ def test_get_template(client, building_motif):
     # Assert
     assert results.status_code == 200
 
-    db_template = building_motif.table_connection.get_db_template_by_id(template.id)
+    db_template = building_motif.table_connection.get_db_template(template.id)
     assert results.json == {
         "id": db_template.id,
         "name": db_template.name,
@@ -77,20 +82,177 @@ def test_get_template_not_found(client):
     assert results.json == {"message": "No template with id -1"}
 
 
-def test_evaluate(client, building_motif):
+def test_evaluate_bindings(client, building_motif):
+    model = Model.create(name="urn:my_model")
     lib = Library.load(directory="tests/unit/fixtures/templates")
     zone = lib.get_template_by_name("zone")
     zone.inline_dependencies()
     assert zone.parameters == {"name", "cav"}
 
     results = client.post(
-        f"/templates/{zone.id}/evaluate",
-        json={"name": {"@id": BLDG["zone1"]}, "cav": {"@id": BLDG["cav1"]}},
+        f"/templates/{zone.id}/evaluate/bindings",
+        json={
+            "model_id": model.id,
+            "bindings": {"name": {"@id": BLDG["zone1"]}, "cav": {"@id": BLDG["cav1"]}},
+        },
     )
 
     assert results.status_code == status.HTTP_200_OK
     graph = Graph().parse(data=results.data, format="ttl")
-    assert (BLDG["cav1"], A, BRICK.CAV) in graph
-    assert (BLDG["zone1"], A, BRICK.HVAC_Zone) in graph
-    assert (BLDG["zone1"], BRICK.isFedBy, BLDG["cav1"]) in graph
+    assert (model.name + "/" + BLDG["cav1"], A, BRICK.CAV) in graph
+    assert (model.name + "/" + BLDG["zone1"], A, BRICK.HVAC_Zone) in graph
+    assert (
+        model.name + "/" + BLDG["zone1"],
+        BRICK.isFedBy,
+        model.name + "/" + BLDG["cav1"],
+    ) in graph
     assert len(list(graph.triples((None, None, None)))) == 3
+
+
+def test_evaluate_bindings_bad_templated_id(client, building_motif):
+    model = Model.create(name="urn:my_model")
+
+    results = client.post(
+        "/templates/-1/evaluate/bindings",
+        json={
+            "model_id": model.id,
+            "bindings": {"name": {"@id": BLDG["zone1"]}, "cav": {"@id": BLDG["cav1"]}},
+        },
+    )
+
+    assert results.status_code == 404
+
+
+def test_evaluate_bindings_no_body(client, building_motif):
+    lib = Library.load(directory="tests/unit/fixtures/templates")
+    zone = lib.get_template_by_name("zone")
+    zone.inline_dependencies()
+    assert zone.parameters == {"name", "cav"}
+
+    results = client.post(f"/templates/{zone.id}/evaluate/bindings")
+
+    assert results.status_code == 400
+
+
+def test_evaluate_bindings_bad_body(client, building_motif):
+    model = Model.create(name="urn:my_model")
+    lib = Library.load(directory="tests/unit/fixtures/templates")
+    zone = lib.get_template_by_name("zone")
+    zone.inline_dependencies()
+    assert zone.parameters == {"name", "cav"}
+
+    results = client.post(
+        f"/templates/{zone.id}/evaluate/bindings",
+        json={
+            # no model
+            "bindings": {"name": {"@id": BLDG["zone1"]}, "cav": {"@id": BLDG["cav1"]}},
+        },
+    )
+
+    assert results.status_code == 400
+
+    results = client.post(
+        f"/templates/{zone.id}/evaluate/bindings",
+        json={
+            "model_id": model.id,
+            # no bindings
+        },
+    )
+
+    assert results.status_code == 400
+
+
+def test_evaluate_bindings_bad_model_id(client, building_motif):
+    lib = Library.load(directory="tests/unit/fixtures/templates")
+    zone = lib.get_template_by_name("zone")
+    zone.inline_dependencies()
+    assert zone.parameters == {"name", "cav"}
+
+    results = client.post(
+        f"/templates/{zone.id}/evaluate/bindings",
+        json={
+            "model_id": -1,
+            "bindings": {"name": {"@id": BLDG["zone1"]}, "cav": {"@id": BLDG["cav1"]}},
+        },
+    )
+
+    assert results.status_code == 404, results.data
+
+
+def test_evaluate_ingress(client, building_motif):
+    # create a 'MODEL' namespace here to scope the entities we create
+    MODEL = Namespace("urn:my_model/")
+    model = Model.create(name=MODEL)
+    lib = Library.load(directory="tests/unit/fixtures/templates")
+    zone = lib.get_template_by_name("zone")
+    zone.inline_dependencies()
+    assert zone.parameters == {"name", "cav"}
+
+    results = client.post(
+        f"/templates/{zone.id}/evaluate/ingress?model_id={model.id}",
+        data=BINDINGS,
+    )
+
+    assert results.status_code == status.HTTP_200_OK, results.data
+    graph = Graph().parse(data=results.data, format="ttl")
+    assert (MODEL["cav1"], A, BRICK.CAV) in graph
+    assert (MODEL["zone1"], A, BRICK.HVAC_Zone) in graph
+    assert (MODEL["zone1"], BRICK.isFedBy, MODEL["cav1"]) in graph
+    assert (MODEL["cav2"], A, BRICK.CAV) in graph
+    assert (MODEL["zone2"], A, BRICK.HVAC_Zone) in graph
+    assert (MODEL["zone2"], BRICK.isFedBy, MODEL["cav2"]) in graph
+
+
+def test_evaluate_ingress_no_body(client, building_motif):
+    lib = Library.load(directory="tests/unit/fixtures/templates")
+    zone = lib.get_template_by_name("zone")
+    zone.inline_dependencies()
+    assert zone.parameters == {"name", "cav"}
+
+    results = client.post(f"/templates/{zone.id}/evaluate/ingress")
+
+    assert results.status_code == 400
+
+
+def test_evaluate_ingress_bad_body(client, building_motif):
+    model = Model.create(name="urn:my_model")
+    lib = Library.load(directory="tests/unit/fixtures/templates")
+    zone = lib.get_template_by_name("zone")
+    zone.inline_dependencies()
+    assert zone.parameters == {"name", "cav"}
+
+    results = client.post(
+        f"/templates/{zone.id}/evaluate/ingress",
+        json={
+            # no model
+            "bindings": {"name": {"@id": BLDG["zone1"]}, "cav": {"@id": BLDG["cav1"]}},
+        },
+    )
+
+    assert results.status_code == 400
+
+    results = client.post(
+        f"/templates/{zone.id}/evaluate/ingress",
+        json={
+            "model_id": model.id,
+            # no bindings
+        },
+    )
+
+    assert results.status_code == 400
+
+
+def test_evaluate_ingress_bad_model_id(client, building_motif):
+    lib = Library.load(directory="tests/unit/fixtures/templates")
+    zone = lib.get_template_by_name("zone")
+    zone.inline_dependencies()
+    assert zone.parameters == {"name", "cav"}
+
+    results = client.post(
+        f"/templates/{zone.id}/evaluate/ingress?model_id=-1",
+        json={
+            "bindings": {"name": {"@id": BLDG["zone1"]}, "cav": {"@id": BLDG["cav1"]}},
+        },
+    )
+
+    assert results.status_code == 404, results.data
